requirements.txt
djangorestframework
django
djangorestframework-simplejwt
djangorestframework-authtoken